+++ ethers-contract/src/multicall/mod.rs
@@ -29,6 +29,8 @@ pub static ADDRESS_BOOK: Lazy<HashMap<U256, Address>> = Lazy::new(|| {
         (Chain::XDai.into(), decode_address("b5b692a88bdfc81ca69dcb1d924f59f0413a602a")),
         (Chain::Polygon.into(), decode_address("11ce4B23bD875D7F5C6a31084f55fDe1e9A87507")),
         (Chain::PolygonMumbai.into(), decode_address("08411ADd0b5AA8ee47563b146743C13b3556c9Cc")),
+        (Chain::Fantom.into(), decode_address("C30EB95BC3ff9D322C4300b65a1575F09b4a3eB1")),
+        (Chain::FantomTestnet.into(), decode_address("280A512EB24Fb655395E0C52D06dcf2dE5253172")),
     ]
     .into()
 });
@@ -187,14 +189,7 @@ impl<M: Middleware> Multicall<M> {
     }
 
     /// Appends a `call` to the list of calls for the Multicall instance
-    ///
-    /// # Panics
-    ///
-    /// If more than the maximum number of supported calls are added. The maximum
-    /// limits is constrained due to tokenization/detokenization support for tuples
     pub fn add_call<D: Detokenize>(&mut self, call: ContractCall<M, D>) -> &mut Self {
-        assert!(self.calls.len() < 16, "Cannot support more than {} calls", 16);
-
         match (call.tx.to(), call.tx.data()) {
             (Some(NameOrAddress::Address(target)), Some(data)) => {
                 let call = Call { target: *target, data: data.clone(), function: call.function };
@@ -284,23 +279,58 @@ impl<M: Middleware> Multicall<M> {
     /// # }
     /// ```
     ///
+    /// # Panics
+    ///
+    /// If more than the maximum number of supported calls are added. The maximum
+    /// limits is constrained due to tokenization/detokenization support for tuples
+    ///
     /// Note: this method _does not_ send a transaction from your account
     ///
     /// [`ContractError<M>`]: crate::ContractError<M>
     pub async fn call<D: Detokenize>(&self) -> Result<D, ContractError<M>> {
-        let contract_call = self.as_contract_call();
+        assert!(self.calls.len() < 16, "Cannot decode more than {} calls", 16);
+        let tokens = self.call_raw().await?;
+        let tokens = vec![Token::Tuple(tokens)];
+        let data = D::from_tokens(tokens)?;
+        Ok(data)
+    }
 
+    /// Queries the Ethereum blockchain via an `eth_call`, but via the Multicall contract and
+    /// without detokenization.
+    ///
+    /// It returns a [`ContractError<M>`] if there is any error in the RPC call.
+    ///
+    /// ```no_run
+    /// # async fn foo() -> Result<(), Box<dyn std::error::Error>> {
+    /// # use ethers_core::types::{U256, Address};
+    /// # use ethers_providers::{Provider, Http};
+    /// # use ethers_contract::Multicall;
+    /// # use std::convert::TryFrom;
+    /// #
+    /// # let client = Provider::<Http>::try_from("http://localhost:8545")?;
+    /// #
+    /// # let multicall = Multicall::new(client, None).await?;
+    /// // The consumer of the API is responsible for detokenizing the results
+    /// // as the results will be a Vec<Token>
+    /// let tokens = multicall.call_raw().await?;
+    /// # Ok(())
+    /// # }
+    /// ```
+    ///
+    /// Note: this method _does not_ send a transaction from your account
+    ///
+    /// [`ContractError<M>`]: crate::ContractError<M>
+
+    pub async fn call_raw(&self) -> Result<Vec<Token>, ContractError<M>> {
+        let contract_call = self.as_contract_call();
         // Fetch response from the Multicall contract
         let (_block_number, return_data) = contract_call.call().await?;
-
-        // Decode return data into ABI tokens
         let tokens = self
             .calls
             .iter()
             .zip(&return_data)
             .map(|(call, bytes)| {
                 let mut tokens: Vec<Token> = call.function.decode_output(bytes.as_ref())?;
-
                 Ok(match tokens.len() {
                     0 => Token::Tuple(vec![]),
                     1 => tokens.remove(0),
@@ -308,14 +338,7 @@ impl<M: Middleware> Multicall<M> {
                 })
             })
             .collect::<Result<Vec<Token>, ContractError<M>>>()?;
-
-        // Form tokens that represent tuples
-        let tokens = vec![Token::Tuple(tokens)];
-
-        // Detokenize from the tokens into the provided tuple D
-        let data = D::from_tokens(tokens)?;
-
-        Ok(data)
+        Ok(tokens)
     }
 
     /// Signs and broadcasts a batch of transactions by using the Multicall contract as proxy.